backend/requirements.txt
Django
django-auth-ldap
djangorestframework
django-cors-headers
djangorestframework-jwt
django-filter
Pillow
drf-friendly-errors
